third_party/blink/web_tests/http/tests/devtools/elements/styles-2/metrics-box-sizing-expected.txt

Tests that content-box and border-box content area dimensions are handled property by the Metrics pane.


Running: testBorderBoxInit1

Running: testInitialBorderBoxMetrics
=== Initial border-box ===
margin: 1 1 1 1
border: 3 3 3 3
padding: 7 7 7 7
content: 35 x 35

Running: testModifiedBorderBoxMetrics
=== Modified border-box ===
margin: 1 1 1 1
border: 3 3 3 3
padding: 7 7 7 7
content: 60 x 35

Running: testContentBoxInit1

Running: testInitialContentBoxMetrics
=== Initial content-box ===
margin: 1 1 1 1
border: 3 3 3 3
padding: 7 7 7 7
content: 55 x 55

Running: testModifiedContentBoxMetrics
=== Modified content-box ===
margin: 1 1 1 1
border: 3 3 3 3
padding: 7 7 7 7
content: 60 x 55
